refactor: move logger to components package
