+++ apps/hobbies-helsinki/src/domain/app/getHobbiesStaticProps.ts
@@ -4,9 +4,9 @@ import type { CmsLanguage } from 'events-helsinki-components';
 import { getMenuLocationFromLanguage } from 'events-helsinki-components';
 import type { GetStaticPropsContext, GetStaticPropsResult } from 'next';
 
+import { staticGenerationLogger } from '../../logger';
 import { getLocaleOrError } from '../../utils/routerUtils';
 import initializeApolloClient from '../clients/eventsFederationApolloClient';
-import { staticGenerationLogger } from '../logger';
 import AppConfig from './AppConfig';
 
 const GLOBAL_QUERY = gql`
