+++ apps/hobbies-helsinki/src/domain/clients/eventsFederationApolloClient.ts
@@ -23,9 +23,9 @@ import {
   sortMenuItems,
 } from 'events-helsinki-components';
 import { useMemo } from 'react';
+import { logger } from '../../logger';
 import { rewriteInternalURLs } from '../../utils/routerUtils';
 import AppConfig from '../app/AppConfig';
-import { logger } from '../logger';
 
 const eventsFederationApolloClient = new MutableReference<
   ApolloClient<NormalizedCacheObject>
