+++ apps/sports-helsinki/package.json
@@ -47,7 +47,6 @@
     "@svgr/webpack": "6.3.1",
     "@types/jsdom": "20.0.0",
     "@types/lodash": "^4.14.182",
-    "chalk": "^4.1.2",
     "classnames": "^2.3.1",
     "copy-to-clipboard": "^3.3.1",
     "cross-fetch": "^3.1.5",
