+++ packages/components/package.json
@@ -39,6 +39,7 @@
     "@storybook/addons": "6.5.12",
     "@storybook/client-api": "6.5.12",
     "@types/react-datepicker": "4.4.2",
+    "chalk": "^4.1.2",
     "classnames": "2.3.2",
     "copy-to-clipboard": "3.3.2",
     "date-fns": "2.29.3",
