+++ packages/components/src/index.ts
@@ -7,3 +7,4 @@ export * from './apollo/utils';
 export * from './constants';
 export * from './types';
 export * from './utils';
+export { default as createLogger } from './loggers/logger';
